[Chore] [GGFE-230] store modal 스토리 폴더 변경
