+++ stories/store/Modal/EditMegaphoneModal.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from '@storybook/react';
 import EditMegaphoneModal from 'components/modal/store/inventory/EditMegaphoneModal';
 
 const meta: Meta<typeof EditMegaphoneModal> = {
-  title: 'InventoryModal/EditMegaphoneModal',
+  title: 'store/Modal/EditMegaphoneModal',
   component: EditMegaphoneModal,
   tags: ['autodocs'],
   argTypes: { receiptId: { control: { type: 'number' } } },
