+++ stories/store/Modal/NewMegaphoneModal.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from '@storybook/react';
 import NewMegaphoneModal from 'components/modal/store/inventory/NewMegaphoneModal';
 
 const meta: Meta<typeof NewMegaphoneModal> = {
-  title: 'InventoryModal/NewMegaphoneModal',
+  title: 'store/Modal/NewMegaphoneModal',
   component: NewMegaphoneModal,
   tags: ['autodocs'],
   argTypes: { receiptId: { control: { type: 'number' } } },
